requirements.txt
torch==2.1.1
torchinfo==1.8.0
torchvision==0.16.1
tqdm==4.66.1
pillow==10.1.0
pyyaml==6.0.1 
pandas==2.1.3